+++ mlxtend/image/eyepad_align.py
@@ -27,15 +27,36 @@ class EyepadAlign():
        2. Transformation is performed based on the eyes' middle point.
 
        3. Finally, the transformed image is padded with zeros to match
-        the desired final image size."""
-    def __init__(self, target_landmarks=None,
-                 image_width=None, image_height=None):
-        self.target_landmarks = target_landmarks
-        self.target_width = image_width
-        self.target_height = image_height
+        the desired final image size.
 
-        self.eyes_mid_point = None
-        self.eyes_distance = None
+    Parameters
+    ----------
+
+    eye_distance:
+
+    verbose :
+
+    Attributes
+    ----------
+
+    target_landmarks_ :
+
+    target_width_ :
+
+    target_height_ :
+
+
+    Examples
+    --------
+
+
+    For more usage examples, please see
+    http://rasbt.github.io/mlxtend/user_guide/image/EyepadAlign/
+
+    """
+    def __init__(self, eye_distance=None, verbose=0):
+        self.eye_distance = eye_distance
+        self.verbose = verbose
 
     def fit(self, target_image=None,
             target_img_dir=None, file_extensions='.jpg'):
@@ -45,51 +66,55 @@ class EyepadAlign():
              b. otherwise, if a target directory is given,
                 calculates the average landmarks for all face images
                 in the directory which will be set as the target landmark.
+
         """
         if target_image is not None:
             landmarks = extract_face_landmarks(target_image)
-            if landmarks is not None:
-                self.target_landmarks = landmarks
-            self.target_width = target_image.shape[1]
-            self.target_height = target_image.shape[0]
+            self.target_landmarks_ = landmarks
+            self.target_width_ = target_image.shape[1]
+            self.target_height_ = target_image.shape[0]
+
         elif target_img_dir is not None:
             file_list = listdir(target_img_dir, file_extensions)
-            print("Fitting the average facial landmarks "
-                  "for {} face images ".format(len(file_list)))
+            if self.verbose >= 1:
+                print("Fitting the average facial landmarks "
+                      "for {} face images ".format(len(file_list)))
             landmarks_list = []
             for f in file_list:
                 img = read_image(filename=f, path=target_img_dir)
                 landmarks = extract_face_landmarks(img)
                 if landmarks is not None:
                     landmarks_list.append(landmarks)
-            self.target_landmarks = np.mean(landmarks_list, axis=0)
-            self.target_width = img.shape[1]
-            self.target_height = img.shape[0]
+            self.target_landmarks_ = np.mean(landmarks_list, axis=0)
+            self.target_width_ = img.shape[1]
+            self.target_height_ = img.shape[0]
 
-    def cal_eye_properties(self, landmarks):
+    def _cal_eye_properties(self, landmarks):
+        """ DOCUMENTATION """
         left_eye = np.mean(landmarks[left_indx], axis=0)
         right_eye = np.mean(landmarks[right_indx], axis=0)
         eyes_mid_point = (left_eye + right_eye)/2.0
-        eyes_distance = np.sqrt(np.sum(np.square(left_eye - right_eye)))
+        eye_distance = np.sqrt(np.sum(np.square(left_eye - right_eye)))
 
-        return eyes_mid_point, eyes_distance
+        return eyes_mid_point, eye_distance
 
     def transform(self, img):
-        if self.eyes_distance is None:
-            props = self.cal_eye_properties(self.target_landmarks)
+        """ DOCUMENTATION """
+        if self.eye_distance is None:
+            props = self._cal_eye_properties(self.target_landmarks_)
             self.eyes_mid_point = props[0]
-            self.eyes_distance = props[1]
+            self.eye_distance = props[1]
         landmarks = extract_face_landmarks(img)
         if landmarks is None:
             return
-        eyes_mid_point, eyes_distance = self.cal_eye_properties(landmarks)
+        eyes_mid_point, eye_distance = self._cal_eye_properties(landmarks)
 
-        scale = self.eyes_distance / eyes_distance
+        scale = self.eye_distance / eye_distance
         tr = (self.eyes_mid_point/scale - eyes_mid_point)
         tr = (int(tr[0]*scale), int(tr[1]*scale))
 
         tform = AffineTransform(scale=(scale, scale), rotation=0, shear=0,
                                 translation=tr)
-        h, w = self.target_height, self.target_width
+        h, w = self.target_height_, self.target_width_
         img_tr = warp(img, tform.inverse, output_shape=(h, w))
         return np.array(img_tr*255, dtype='uint8')